requirements.txt
fastapi
uvicorn[standard]
httpx
gspread
oauth2client
python-dotenv
